+++ boards/neo_trinkey/src/lib.rs
@@ -4,7 +4,7 @@
 pub use cortex_m_rt::entry;
 
 pub use atsamd_hal as hal;
-pub use embedded_hal as ehal;
+pub use hal::ehal;
 pub use hal::pac;
 
 #[cfg(feature = "usb")]