+++ boards/trellis_m4/examples/neopixel_keypad.rs
@@ -9,8 +9,8 @@ use panic_semihosting as _;
 use trellis_m4 as bsp;
 use ws2812_timer_delay as ws2812;
 
-use embedded_hal::digital::v1_compat::OldOutputPin;
-use embedded_hal::digital::v2::InputPin;
+use hal::ehal::digital::v1_compat::OldOutputPin;
+use hal::ehal::digital::v2::InputPin;
 
 use bsp::entry;
 use hal::pac::{CorePeripherals, Peripherals};