+++ boards/trellis_m4/examples/neopixel_orientation.rs
@@ -11,7 +11,7 @@ use panic_semihosting as _;
 use trellis_m4 as bsp;
 use ws2812_timer_delay as ws2812;
 
-use embedded_hal::digital::v1_compat::OldOutputPin;
+use hal::ehal::digital::v1_compat::OldOutputPin;
 
 use bsp::entry;
 use hal::adxl343::accelerometer::Orientation;