+++ boards/samd11_bare/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- remove extraneous `embedded-hal` dependencies from BSPs
 * move `usbd-x` crates used only in examples to `[dev-dependencies]`
 * Bump `cortex-m`/`cortex-m-rt` dependencies to fix a build issue