+++ boards/wio_lite_mg126/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- remove extraneous `embedded-hal` dependencies from BSPs
 - move `usbd-x` crates used only in examples to `[dev-dependencies]`
 - removed unnecessary dependency on `nb` and `panic_rtt` (#510)